test(arenabench): split the cloud CLI witnesses out of test_cloud.py
Merging main brought #3279's four command-surface test classes alongside
this branch's TestBalancePreflight, and the combined file hit 1614 lines —
over the 1500-line ceiling, not grandfathered. The ratchet's instruction is
to split, not to add a baseline entry, and the file already had a clean
seam at the 'the CLI path' banner.

  cloud_fakes.py          the recorded-call AWS fakes, the executor factory
                          and the match template both halves build on
  test_cloud.py           the submit contract and the pure decisions
  test_cloud_commands.py  the _cmd_cloud_* surface: what a run refuses
                          before it spends, and fetch/status/cancel

The fakes moved to a helper module rather than being duplicated or
imported test-module-to-test-module; tests/ has no __init__.py, so pytest
puts it on sys.path and 'from cloud_fakes import ...' resolves. The naming
follows the test_cloud_merge.py / test_cloud_watch.py siblings already
there.

Pure move: the 104 collected class and test names are identical before and
after (75 tests, 59 + 16). check-file-size.sh passes.
